Mv fns to account, buffer, hash, signature and object files
